pyproject.toml
[tool.poetry]
name = "hunknote"
version = "1.0.0"
description = "AI-powered git commit message generator with multi-LLM support"
authors = ["Avinash Ranganath <nash911@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/hunknote/hunknote"
repository = "https://github.com/hunknote/hunknote"
documentation = "https://github.com/hunknote/hunknote#readme"
keywords = ["git", "commit", "ai", "llm", "cli", "anthropic", "openai", "gemini", "commit-message", "hunknote"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
packages = [{include = "hunknote"}]
include = ["README.md", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.21.0"
pydantic = "^2.5.0"
python-dotenv = "^1.2.1"
pyyaml = "^6.0.0"
# LLM Provider SDKs
anthropic = "^0.78.0"
openai = "^1.0.0"
google-genai = "^1.0.0"
mistralai = "^1.0.0"
cohere = "^5.0.0"
groq = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.12.0"

[tool.poetry.scripts]
hunknote = "hunknote.cli:app"
git-hunknote = "hunknote.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
